+++ src/novelystlib/view_controller/novelyst_tk.py
@@ -342,12 +342,12 @@ class NovelystTk(MainTk):
         self.exportMenu.add_separator()
         self.exportMenu.add_command(label=_('Manuscript without tags (export only)'), command=lambda: self.export_document('', lock=False))
         self.exportMenu.add_command(label=_('Brief synopsis (export only)'), command=lambda: self.export_document('_brf_synopsis', lock=False))
-        self.exportMenu.add_command(label=_('Arcs (export only)'), command=lambda: self.export_document('_arcs', lock=False))
         self.exportMenu.add_command(label=_('Cross references (export only)'), command=lambda: self.export_document('_xref', lock=False))
         self.exportMenu.add_separator()
         self.exportMenu.add_command(label=_('Obfuscated text for word count'), command=lambda: self.export_document('_wrimo', lock=False, show=False))
         self.exportMenu.add_command(label=_('Characters/locations/items data files'), command=lambda: self.export_document('_data', lock=False, show=False))
         self.exportMenu.add_separator()
+        self.exportMenu.add_command(label=_('Plot description (export only)'), command=lambda: self.export_document('_plot', lock=False))
         self.exportMenu.add_command(label=_('Plot spreadsheet (export only)'), command=lambda: self.export_document('_plotlist', lock=False))
         self.exportMenu.add_command(label=_('Show Plot list'), command=lambda: self._show_report('_plotlist'))